third_party/WebKit/LayoutTests/http/tests/security/xss-DENIED-window-index-assign-expected.txt

ALERT: PASS: Access to parent[0].a threw exception.
ALERT: PASS: Access to parent[0].f() threw exception.

This test passes if the access is forbidden.
